+++ packages/jira/src/matrics/get-sprint-variance.ts
@@ -69,7 +69,7 @@ export async function sprintHitsResponse(
           esb.termQuery('body.projectId', projectId),
           esb.termQuery('body.isDeleted', false),
           esb.boolQuery().should(dateRangeQueries).minimumShouldMatch(1),
-          esb.termQuery('body.state', state),
+          esb.termsQuery('body.state', state.split(',')),
         ])
     )
     .sort(esb.sort('body.startDate', 'desc'))
@@ -92,7 +92,7 @@ export async function versionHitsResponse(
   projectId: string,
   dateRangeQuery: esb.RangeQuery[],
   reqCtx: Other.Type.RequestCtx,
-  versionState?: string
+  versionState: string
 ): Promise<{
   versionHits: [] | (Pick<Other.Type.Hit, '_id'> & Other.Type.HitBody)[];
   totalPages: number;
@@ -108,7 +108,7 @@ export async function versionHitsResponse(
           esb.termQuery('body.projectId', projectId),
           esb.termQuery('body.isDeleted', false),
           esb.boolQuery().should(dateRangeQuery).minimumShouldMatch(1),
-          esb.termQuery('body.status', versionState),
+          esb.termsQuery('body.status', versionState.split(',')),
         ])
     )
     .sort(esb.sort('body.startDate', 'desc'))
@@ -620,16 +620,7 @@ export function createVersionQuery(
           esb.termQuery('body.projectId', projectId),
           esb.termQuery('body.isDeleted', false),
           esb.boolQuery().should(dateRange).minimumShouldMatch(1),
-          esb
-            .boolQuery()
-            .must(
-              state
-                ? esb.termQuery('body.status', state)
-                : esb.termsQuery('body.status', [
-                    Jira.Enums.State.RELEASED,
-                    Jira.Enums.State.UNRELEASED,
-                  ])
-            ),
+          esb.boolQuery().must(esb.termsQuery('body.status', state)),
         ])
     )
     .sort(esb.sort('body.startDate', 'desc'));
@@ -750,13 +741,7 @@ export function createSprintQuery(
           esb.termQuery('body.projectId', projectId),
           esb.termQuery('body.isDeleted', false),
           esb.boolQuery().should(dateRangeQueries).minimumShouldMatch(1),
-          esb
-            .boolQuery()
-            .must(
-              state
-                ? esb.termQuery('body.state', state)
-                : esb.termsQuery('body.state', [Jira.Enums.State.CLOSED, Jira.Enums.State.ACTIVE])
-            ),
+          esb.boolQuery().must(esb.termsQuery('body.state', state)),
         ])
     )
     .sort(esb.sort('body.sprintId'));