+++ CHANGELOG.md
@@ -1,5 +1,23 @@
 # Changelog
 
+## [1.22.3](https://github.com/studiographene/pulse-data-integration/compare/v1.22.2...v1.22.3) (2025-03-27)
+
+
+### Bug Fixes
+
+* **PX-2334:** fix average ([be9d7fd](https://github.com/studiographene/pulse-data-integration/commit/be9d7fd926fa8771a3a3a1bd2fda11a1c7594fc1))
+* **PX-2334:** fix average ([#1125](https://github.com/studiographene/pulse-data-integration/issues/1125)) ([a95e339](https://github.com/studiographene/pulse-data-integration/commit/a95e339de70befcde6ed8859839163772297ea4d))
+
+## [1.22.2](https://github.com/studiographene/pulse-data-integration/compare/v1.22.1...v1.22.2) (2025-03-27)
+
+
+### Bug Fixes
+
+* **PX-2551:** sprint and version default state ([9e54227](https://github.com/studiographene/pulse-data-integration/commit/9e5422744251233ad78db92889498b27c8634579))
+* **PX-2551:** sprint and version default state ([#1123](https://github.com/studiographene/pulse-data-integration/issues/1123)) ([769989e](https://github.com/studiographene/pulse-data-integration/commit/769989e3b08f55f5e03418297a0d688762afc142))
+* **PX-2551:** sprint state ([2c9f019](https://github.com/studiographene/pulse-data-integration/commit/2c9f019558e193f0ccde566be176f4141fcdf379))
+* **PX-2551:** sprint state ([#1120](https://github.com/studiographene/pulse-data-integration/issues/1120)) ([dc03dd6](https://github.com/studiographene/pulse-data-integration/commit/dc03dd6dae3acc9c5222cc443d8fc11d86bf0c8d))
+
 ## [1.22.1](https://github.com/studiographene/pulse-data-integration/compare/v1.22.0...v1.22.1) (2025-03-21)